Khepri: Project the topic trie graph based on rabbit_db_bindings
When using mnesia as the backing metadata store, the topic trie graph
used for topic exchange routing is constructed using the
'rabbit_exchange_type' callbacks on 'rabbit_exchange_type_topic'.

The extra topic graph tables are difficult to migrate directly to
Khepri because of the graph layout and migration took a considerable
amount of time when many bindings were present. Instead, we can
construct the topic trie graph in a Khepri projection of the bindings
stored in the main bindings subtree under 'rabbit_db_binding'. The
migration for regular bindings takes linear time on the number of
bindings so this is very fast.

The projection for the topic trie in Khepri only needs the additional
knowledge of the type of the exchange a binding is attached to. We can
use a `#if_data_matches{}` query on the exchange component of the
path in the tree to match only the bindings of topic exchanges. When
creating bindings we already look up the exchange so this information
is already available and only needs an extra 'put'.

This change removes the rabbit_exchange_type callbacks when using
Khepri and refactors the topic trie projection to work on the bindings
under rabbit_db_binding rather than extra storage in the tree.